Revert "fixed #46 delete, update API"

+++ controllers/todos.js
@@ -3,21 +3,19 @@ const { Op } = require('sequelize');
 const Todos = models.todo;
 
 module.exports = {
-  get: async (req, res) => {
-    console.log(req.params)
+  get: (req, res) => {
     if (req.session.userId) {
       if (req.params.yearMonth) {
-        console.log('hasparams');
-        await Todos.findAll({
+        Todos.findAll({
           where: {
             userId: req.session.userId,
             startDate: { [Op.startsWith]: req.params.yearMonth },
           },
         })
-          .then((data) => console.log(data)) //res.status(200).send(data)
+          .then((data) => res.status(200).send(data))
           .catch(() => res.status(404).send({ error: '404 Not Found' }));
       } else {
-        await Todos.findAll({ where: { userId: req.session.userId } })
+        Todos.findAll({ where: { userId: req.session.userId } })
           .then((data) => res.status(200).send(data))
           .catch(() => res.status(404).send({ error: '404 Not Found' }));
       }
@@ -25,11 +23,11 @@ module.exports = {
       res.status(400).send({ error: '400 Bad Request' });
     }
   },
-  post: async (req, res) => {
+  post: (req, res) => {
     console.log('req.session', req.session);
     console.log('req.body', req.body);
     if (req.session.userId) {
-      await Todos.create({
+      Todos.create({
         userId: req.session.userId,
         title: req.body.title,
         body: req.body.body,
@@ -43,30 +41,18 @@ module.exports = {
       res.status(400).send({ error: '저기400 Bad Request' });
     }
   },
-  update: async (req, res) => {
-    console.log(req.params);
+  update: (req, res) => {
     if (req.session.userId) {
-      await Todos.update(
+      Todos.update(
         {
           title: req.body.title,
           body: req.body.body,
           startDate: req.body.startDate,
           endDate: req.body.endDate,
         },
-        {
-          where: {
-            userId: req.session.userId,
-            id: req.params.id
-          }
-        }
+        { where: { id: req.session.userId } }
       )
-
-      await Todos.findOne({
-        where: {
-          userId: req.session.userId,
-          id: req.params.id
-        }
-      }).then((data) => res.status(200).send(data))
+        .then((data) => res.status(200).send(data))
         .catch(() => res.status(404).send({ error: '404 Not Found' }));
     } else {
       res.status(400).send({ error: '400 Bad Request' });
@@ -74,12 +60,7 @@ module.exports = {
   },
   delete: (req, res) => {
     if (req.session.userId) {
-      Todos.destroy({
-        where: {
-          userId: req.session.userId,
-          id: req.params.id
-        }
-      })
+      Todos.destroy({ where: { id: req.session.userId } })
         .then(() => res.status(200).send({ result: 'Success' }))
         .catch(() => res.status(400).send({ error: '400 Bad Request' }));
     } else {
